new styles added

+++ login2_system.py
@@ -84,14 +84,12 @@ class Login_system:
             self.txt_password.config(show="")
 
     def iExit_2(self):
-
-        self.iExit_2 = messagebox.askyesno("Password Manager Window", "Confirm if you want to Logout.", parent=self.root2)
-        if self.iExit_2 > 0:
+        MsgBox = messagebox.askquestion('Exit Application', 'Are you sure you want to Logout the application', icon='warning',parent=self.root2)
+        if MsgBox == 'yes':
             self.root2.destroy()
-
         else:
-            command = self.root2
-            return
+            pass
+
 
 
 
@@ -157,6 +155,15 @@ class Login_system:
                 con.close()
             except Exception as es:
                     messagebox.showerror("Error",f"Error Due to: {str(es)}",parent=self.root)
+
+    """def delete_account(self):
+        if self.txt_username.get()=="" or self.txt_password.get()=="":
+            messagebox.showerror("Error","All fields are required",parent=self.root)
+        else:
+            try:
+                db=pymysql.connect("localhost","root",'',"password_database")
+                cur=db.cursor()
+                cur.execute()"""
     def password_manager_window(self):
         self.root2=Toplevel(self.root)
         self.root2.title("Password manager Window")
@@ -176,7 +183,7 @@ class Login_system:
         self.Username_var = StringVar()
         self.URL_var = StringVar()
         self.Password_var = StringVar()
-        self.Notes_var = StringVar()
+        self.EmailID_var = StringVar()
         self.U_ID_var=StringVar()
         self.search_by=StringVar()
         self.search_txt=StringVar()
@@ -200,8 +207,8 @@ class Login_system:
         self.txt_URL=Entry(Manager_Frame,textvariable=self.URL_var,font=("times new roman",20,"bold"),bd=5,relief=GROOVE).place(x=140,y=180)
         lbl_password=Label(Manager_Frame, text="Password", bg="white", fg="black", font=("times new roman", 20, "bold")).place(x=10,y=230)
         self.txt_password_entry=Entry(Manager_Frame,textvariable=self.Password_var,font=("times new roman",20,"bold"),bd=5,relief=GROOVE).place(x=140,y=230)
-        lbl_notes=Label(Manager_Frame, text="Notes", bg="white", fg="black", font=("times new roman", 20, "bold")).place(x=10,y=280)
-        self.txt_notes = Entry(Manager_Frame,textvariable=self.Notes_var, font=("times new roman", 20, "bold"), bd=5, relief=GROOVE).place(x=140,y=280,height=100)
+        lbl_EmailID=Label(Manager_Frame, text="Email ID", bg="white", fg="black", font=("times new roman", 20, "bold")).place(x=10,y=280)
+        self.txt_EmailID = Entry(Manager_Frame,textvariable=self.EmailID_var, font=("times new roman", 20, "bold"), bd=5, relief=GROOVE).place(x=140,y=280)
         self.txt_U_ID=Entry(Manager_Frame,textvariable=self.U_ID_var,font=("times new roman",20,"bold"),bd=5,relief=GROOVE,state="readonly",highlightbackground="red",highlightthickness=2).place(x=10,y=390,width=120)
 
 
@@ -231,7 +238,11 @@ class Login_system:
 
         scroll_x=ttk.Scrollbar(Table_Frame,orient=HORIZONTAL)
         scroll_y = ttk.Scrollbar(Table_Frame, orient=VERTICAL)
-        self.Password_Table=ttk.Treeview(Table_Frame,columns=("Title","Username","URL","Password","Notes","U_ID"),xscrollcommand=scroll_x.set,yscrollcommand=scroll_y.set)
+        self.style=ttk.Style()
+        self.style.theme_use("clam")
+        self.style.configure("Treeview",background="grey71",foreground="black",rowheight=25,fieldbackground="grey71")
+        self.style.map("Treeview",background=[("selected","green")])
+        self.Password_Table=ttk.Treeview(Table_Frame,columns=("Title","Username","URL","Password","Email_ID","U_ID"),xscrollcommand=scroll_x.set,yscrollcommand=scroll_y.set)
 
 
         
@@ -243,7 +254,7 @@ class Login_system:
         self.Password_Table.heading("Username", text="Username")
         self.Password_Table.heading("URL", text="URL")
         self.Password_Table.heading("Password", text="Password")
-        self.Password_Table.heading("Notes", text="Notes")
+        self.Password_Table.heading("Email_ID", text="Email_ID")
         self.Password_Table.heading("U_ID", text="U_ID")
 
         self.Password_Table["show"]="headings"
@@ -251,37 +262,41 @@ class Login_system:
         self.Password_Table.column("Username", width=200)
         self.Password_Table.column("URL", width=200)
         self.Password_Table.column("Password", width=200)
-        self.Password_Table.column("Notes", width=200)
+        self.Password_Table.column("Email_ID", width=200)
         self.Password_Table.column("U_ID",width=200)
         self.Password_Table.pack(fill=BOTH,expand=1)
         self.Password_Table.bind("<ButtonRelease-1>",self.get_cursor)
         self.fetch_data()
+
     #======Add entries=======
     def add_data(self):
-        if self.Title_var.get()=="" or self.Username_var.get()=="" or self.URL_var.get()=="" or self.Password_var.get()=="" or self.Notes_var.get()=="":
+        if self.Title_var.get()=="" or self.Username_var.get()=="" or self.URL_var.get()=="" or self.Password_var.get()=="" or self.EmailID_var.get()=="":
             messagebox.showerror("Error","All entries are required to fulfill!" ,parent=self.root2)
         else:
 
             db=pymysql.connect("localhost","root",'',"password_database")
             cur=db.cursor()
-            cur.execute("insert into user_{} (Title,Username,URL,Password,Notes,U_ID) values(%s,%s,%s,%s,%s,%s)".format(str(self.iv_and_salt[2])),(self.Title_var.get(),self.Username_var.get(),
+            cur.execute("insert into user_{} (Title,Username,URL,Password,Email_ID,U_ID) values(%s,%s,%s,%s,%s,%s)".format(str(self.iv_and_salt[2])),(self.Title_var.get(),self.Username_var.get(),
                      self.URL_var.get(),
                      encrypt(self.Password_var.get(),
                              self.security_key,int(self.iv_and_salt[0])),
-                     self.Notes_var.get(),random_ID_generator()))
+                     self.EmailID_var.get(),random_ID_generator()))
             db.commit()
             self.fetch_data()
             self.Clear()
             db.close()
     #======fetch data====
     def fetch_data(self):
+
         db = pymysql.connect("localhost", "root", '', "password_database")
         cur = db.cursor()
         cur.execute("select * from user_{}".format(str(self.iv_and_salt[2])))
         rows=cur.fetchall()
+        self.count=0
         if len(rows)!=0:
             self.Password_Table.delete(*self.Password_Table.get_children())
             for row in rows:
+
                 self.Password_Table.insert("",END,values=(row[0],row[1],row[2],decrypt(binascii.unhexlify(row[3]),self.security_key,int(self.iv_and_salt[0])),row[4],row[5]))
             db.commit()
         db.close()
@@ -291,7 +306,7 @@ class Login_system:
         self.Username_var.set("")
         self.URL_var.set("")
         self.Password_var.set("")
-        self.Notes_var.set("")
+        self.EmailID_var.set("")
         self.U_ID_var.set("")
         self.search_txt.set("")
         self.search_by.set("")
@@ -303,16 +318,16 @@ class Login_system:
         self.Username_var.set(row[1])
         self.URL_var.set(row[2])
         self.Password_var.set(row[3])
-        self.Notes_var.set(row[4])
+        self.EmailID_var.set(row[4])
         self.U_ID_var.set(row[5])
     #=======update data from table=======
 
     def update_data(self):
         db = pymysql.connect("localhost", "root", '', "password_database")
         cur = db.cursor()
-        cur.execute("update user_{} set Title=%s,Username=%s,URL=%s,Password=%s,Notes=%s where U_ID=%s".format(str(self.iv_and_salt[2])),
+        cur.execute("update user_{} set Title=%s,Username=%s,URL=%s,Password=%s,Email_ID=%s where U_ID=%s".format(str(self.iv_and_salt[2])),
                     (self.Title_var.get(),self.Username_var.get(),self.URL_var.get(),encrypt(self.Password_var.get(),
-                             self.security_key,int(self.iv_and_salt[0])),self.Notes_var.get(),self.U_ID_var.get()))
+                             self.security_key,int(self.iv_and_salt[0])),self.EmailID_var.get(),self.U_ID_var.get()))
 
         db.commit()
         self.fetch_data()
@@ -325,24 +340,29 @@ class Login_system:
         cur.execute("delete from user_{} where U_ID=%s".format(str(self.iv_and_salt[2])),(self.U_ID_var.get()))
         db.commit()
         db.close()
+        messagebox.showinfo("Deletion","Selected record has been deleted",parent=self.root2)
         self.fetch_data()
         self.Clear()
 
     def search_option(self):
         db = pymysql.connect("localhost", "root", '', "password_database")
         cur = db.cursor()
-        table_name="select * from user_"+str(self.iv_and_salt[2])
-        cur.execute(table_name+" where " + str(self.search_by.get()) +" Like '%"+str(self.search_txt.get())+"%'")
-        rows = cur.fetchall()
-        if len(rows) != 0:
-            self.Password_Table.delete(*self.Password_Table.get_children())
-            for row in rows:
-                self.Password_Table.insert("", END, values=(row[0], row[1], row[2],
+        if str(self.search_by.get())=="":
+            messagebox.showerror("Error","Empty parameter given ,Please choose another paramter!",parent=self.root2)
+        else:
+
+            table_name="select * from user_"+str(self.iv_and_salt[2])
+            cur.execute(table_name+" where " + str(self.search_by.get()) +" Like '%"+str(self.search_txt.get())+"%'")
+            rows = cur.fetchall()
+            if len(rows) != 0:
+                self.Password_Table.delete(*self.Password_Table.get_children())
+                for row in rows:
+                    self.Password_Table.insert("", END, values=(row[0], row[1], row[2],
                                                             decrypt(binascii.unhexlify(row[3]), self.security_key,
                                                                     int(self.iv_and_salt[0])), row[4], row[5]))
-            db.commit()
-        else:
-            messagebox.showerror("Error","Incorrect Information!",parent=self.root2)
+                db.commit()
+            else:
+                messagebox.showerror("Error","Incorrect Information!",parent=self.root2)
         db.close()
         self.Clear()
     def generate_random_password(self):
